add trigger to report sentry error to electron bridge
