+++ src/background/useTrayService/utils/getBaseMenuItems.ts
@@ -1,10 +1,13 @@
-import * as Sentry from '@sentry/electron/main';
 import { MenuItemConstructorOptions, dialog } from 'electron';
 import log from 'electron-log';
-import { v4 as uuidv4 } from 'uuid';
 
 import { State } from '../../types';
-import { isLinux, isProduction, quitApp } from '../../utils';
+import {
+  isLinux,
+  isProduction,
+  quitApp,
+  triggerSentryError,
+} from '../../utils';
 
 export default (state: State): Array<MenuItemConstructorOptions> => {
   const menuItems: Array<MenuItemConstructorOptions> = [];
@@ -32,16 +35,7 @@ export default (state: State): Array<MenuItemConstructorOptions> => {
       log.info(
         `Detected click on Send Bug Report menu item; sending Sentry alert`
       );
-      Sentry.withScope((scope) => {
-        // Use a unique fingerprint to avoid grouping in Sentry so that we
-        // don't accidentally miss an alert about a manual bug report
-        scope.setFingerprint([uuidv4()]);
-
-        const userEmail = scope.getUser()?.email || `no user`;
-        const now = new Date().toISOString();
-        const message = `Manual bug report - ${userEmail} - ${now}`;
-        Sentry.captureException(new Error(message));
-      });
+      triggerSentryError();
       dialog.showErrorBox(
         `Bug report submitted`,
         `Oh snap! Sorry about that. We'll look into this. Please send us a brief description and screenshot via your company's Slack Connect channel or email us at support@swivvel.io. We'll get back to you ASAP.`
