+++ src/background/utils/index.ts
@@ -16,6 +16,7 @@ import removeQueryParams from './removeQueryParams';
 import shouldOpenUrlInBrowser from './shouldOpenUrlInBrowser';
 import showErrorMessage from './showErrorMessage';
 import sleep from './sleep';
+import triggerSentryError from './triggerSentryError';
 
 export {
   getBrowserWindowName,
@@ -36,4 +37,5 @@ export {
   shouldOpenUrlInBrowser,
   showErrorMessage,
   sleep,
+  triggerSentryError,
 };
